[queue] [api] [refactor] Queue has a new, more robust api
Each queue method takes an id instead of a queue object. Each call to
modify the queue does a get and an update, and calls back with the
result and the queue object. All of these calls are enqueued so that
they execute in-order. Between these changes, the queue should remain
consistent, and be able to recover from changes made to the queue
outside the context of the persistence api.